Add PASTED_NAME_EDITS configuration for copynauto
Allows you to automatically remove, via regexp substitution, information that you wanted in the buffer name
but don't want in the layer name.

By default, remove anything inside double braces : [[ like this ]] , including the braces themselves,
and strip trailing whitespace.

Also add 'isize' and 'where' templating variables.
`where' resolves to the equivalent of '[[{imagewidth}x{imageheight}+{offsetx},{offsety}]]' .

If a `where` spec is detected in the name of a buffer, and the image that is being pasted into matches the dimensions of the source image,
the pasted layer will be moved to the specified offset when pasting via Paste Latest Named and Remove.