copynaut: add new "Export Clippings from Vectors" command.
This loops over the vectors on the image, converting each to a selection and then exporting it (like Export Clipping)
using the vector name as the 'suffix'.

As long as vectors satisfy your needs for defining each selection mask, this can streamline the process of extracting elements from a picture into:
 1. create vectors, naming them appropriately
 2. invoke "Export Clippings from Vectors"

compared to the previous
 1. Create vector
 2. Convert to selection
 3. Export clipping (and type in name)
 4. Go to step 1, until no more regions remain to be extracted.

Definitely more streamlined.

If vectors turn out to not be enough, I may expand this to also support channels, which would work in the same way but allow all of the other types of selections that
vectors are not capable of (partially selected areas, etc)