copynaut: New 'create vectors from grid' command.
Creates a vector for each grid cell that is
a) within the selection, b) not empty (all pixels are same color) and c) not duplicate.
In the event of duplicate cells, the first cell (looking left to right, top to bottom) gets a vector, the
subsequent dupes of it are ignored.

Intended to help extracting tiles and sets of tiles from a collected mess of tiles.
The expected workflow is :

 * Select the tiles you care about (not worrying about dupes or blanks possibly being included)
 * Run this command
 * Rename the vectors accordingly, and in some cases, use 'merge visible vectors' to combine single tiles into a single output region
 * Run 'export clippings from vectors' to generate all the output images.