New Layer (quick) command added to backgroundify.py.
This command creates an image-sized layer with the specified name, mode, and opacity.
For example, when inking a bunch of drawings, I just fill in name=Ink and mode=Multiply,
then I can just hit Enter when it comes up again.

It can be found at 'Layer->Stack->New Layer (quick)' in the menus.

Created to fix three main deficiencies of the New Layer dialog:
  * I don't care about the dimensions. Let them be image sized, I'll crop the layer if memory demands require it.
  * Doesn't allow you to set the mode in the same operation
  * Not non-interactively repeatable (via F2)

In the future, should also fix the following deficiency of the New Layer Dialog
  * Initial fill value is often strongly implied by mode
    (eg. Multiply or Darken Only: you probably want white.
    Add, Sub, Lighten Only or Screen, you probably want black.
    Grain merge/extract, soft or hard light or overlay,
    you probably want (128,128,128) grey.)

This should be compatible with both 2.8 and 2.9. However, it hasn't been tested with 2.9 yet.